requirements_enhanced.txt
# Guardia AI Enhanced System - Complete Dependencies
# Updated with exact versions and comprehensive package list

# =============================================================================
# CORE COMPUTER VISION & AI/ML
# =============================================================================

# Deep Learning & Neural Networks
tensorflow==2.15.0
tensorflow-hub==0.16.1
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.206  # YOLO v8 for object detection
transformers==4.35.2  # Hugging Face transformers
onnxruntime==1.16.3

# Computer Vision
opencv-python==4.8.1.78
pillow==10.0.1
scikit-image==0.22.0
imageio==2.33.0
albumentations==1.3.1
imgaug==0.4.0

# Face Recognition & Detection
face-recognition==1.3.0
dlib==19.24.2
mediapipe==0.10.7
mtcnn==0.1.1
facenet-pytorch==2.5.3
insightface==0.7.3

# Mathematical & Scientific Computing
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0

# =============================================================================
# GOOGLE CLOUD AI SERVICES
# =============================================================================

google-cloud-videointelligence==2.11.1
google-cloud-vision==3.4.5
google-cloud-storage==2.10.0
google-cloud-automl==2.11.1
google-cloud-aiplatform==1.38.1
google-cloud-translate==3.12.1
google-cloud-speech==2.21.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# =============================================================================
# WEB FRAMEWORK & API
# =============================================================================

# FastAPI & Related
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
starlette==0.27.0

# WebSocket & Real-time Communication
websockets==12.0
python-socketio==5.9.0
python-socketio[asyncio_client]==5.9.0

# HTTP Clients & Tools
requests==2.31.0
aiohttp==3.9.0
httpx==0.25.2

# =============================================================================
# DATABASE & STORAGE
# =============================================================================

# MongoDB
pymongo==4.5.0
motor==3.3.2  # Async MongoDB driver
mongoengine==0.27.0

# Redis
redis==5.0.1
aioredis==2.0.1

# SQL Databases (Optional)
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # PostgreSQL async driver
psycopg2-binary==2.9.9  # PostgreSQL sync driver

# =============================================================================
# AUTHENTICATION & SECURITY
# =============================================================================

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
pyjwt==2.8.0
bcrypt==4.0.1

# =============================================================================
# BACKGROUND TASKS & SCHEDULING
# =============================================================================

celery==5.3.4
kombu==5.3.4
schedule==1.2.0
apscheduler==3.10.4
rq==1.15.1  # Simple job queue

# =============================================================================
# NOTIFICATIONS & ALERTS
# =============================================================================

# Desktop Notifications
plyer==2.1.0

# SMS & Phone
twilio==8.10.0

# Email
sendgrid==6.10.0
yagmail==0.15.293
smtplib2==0.2.1

# Push Notifications
pyfcm==1.5.4
pywhatkit==5.4

# =============================================================================
# VIDEO & MEDIA PROCESSING
# =============================================================================

ffmpeg-python==0.2.0
moviepy==1.0.3
imageio-ffmpeg==0.4.9
av==10.0.0  # PyAV for video processing

# =============================================================================
# UTILITY LIBRARIES
# =============================================================================

# Environment & Configuration
python-dotenv==1.0.0
python-decouple==3.8
pyyaml==6.0.1
configparser==6.0.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3
arrow==1.3.0

# CLI & User Interface
click==8.1.7
rich==13.7.0
typer==0.9.0

# Progress & Logging
tqdm==4.66.1
loguru==0.7.2
structlog==23.2.0

# Async & Concurrency
asyncio==3.4.3
aiofiles==23.2.1
anyio==3.7.1

# =============================================================================
# SERIALIZATION & DATA FORMATS
# =============================================================================

orjson==3.9.10
msgpack==1.0.7
ujson==5.8.0
xmltodict==0.13.0

# =============================================================================
# MONITORING & PERFORMANCE
# =============================================================================

# System Monitoring
psutil==5.9.6
gputil==1.4.0

# Performance Profiling
memory-profiler==0.61.0
line-profiler==4.1.1
py-spy==0.3.14
cprofile-dev==1.0.2

# Application Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# =============================================================================
# TESTING & DEVELOPMENT
# =============================================================================

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.3.1
pytest-mock==3.12.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
bandit==1.7.5

# Development Tools
pre-commit==3.5.0
ipython==8.17.2
jupyter==1.0.0
notebook==7.0.6

# =============================================================================
# TYPE HINTS & ANNOTATIONS
# =============================================================================

typing-extensions==4.8.0
types-requests==2.31.0.10
types-redis==4.6.0.11

# =============================================================================
# ADDITIONAL SPECIALIZED LIBRARIES
# =============================================================================

# Audio Processing (for multi-modal AI)
librosa==0.10.1
soundfile==0.12.1
pyaudio==0.2.11

# Network & Security
paramiko==3.3.1  # SSH client
cryptography==41.0.7
keyring==24.3.0

# File Handling
pathlib2==2.3.7
watchdog==3.0.0
send2trash==1.8.2

# Caching
diskcache==5.6.3
joblib==1.3.2

# Misc Utilities
validators==0.22.0
humanize==4.8.0
colorama==0.4.6
